beautified code.

+++ src/weight.rs
@@ -845,17 +845,17 @@ impl Weight {
     unsafe fn expmx_ps_simd256(x4 : x86_64::__m256) -> x86_64::__m256 {
         // let x4 = x86_64::_mm_load_ps(x);
         // clip x
-        let max4 = x86_64::_mm256_set1_ps(EXP_HI);
+        let max4 = x86_64::_mm256_set1_ps(EXP_HI as f32);
         let x4 = x86_64::_mm256_min_ps(x4, max4);
-        let min4 = x86_64::_mm256_set1_ps(EXP_LO);
+        let min4 = x86_64::_mm256_set1_ps(EXP_LO as f32);
         let x4 = x86_64::_mm256_max_ps(x4, min4);
         let m1 = x86_64::_mm256_set1_ps(-1.0);
         let x4 = x86_64::_mm256_mul_ps(x4, m1);
 
         /* express exp(x) as exp(g + n*log(2)) */
-        let log2ef = x86_64::_mm256_set1_ps(CEPHES_LOG2EF);
+        let log2ef = x86_64::_mm256_set1_ps(CEPHES_LOG2EF as f32);
         let fx = x86_64::_mm256_mul_ps(x4, log2ef);
-        let zp5 = x86_64::_mm256_set1_ps(CEPHES_EXP_P5);
+        let zp5 = x86_64::_mm256_set1_ps(CEPHES_EXP_P5 as f32);
         let fx = x86_64::_mm256_add_ps(fx, zp5);
         let emm0 = x86_64::_mm256_cvtps_epi32(fx);
         let tmp = x86_64::_mm256_cvtepi32_ps(emm0);
@@ -865,30 +865,30 @@ impl Weight {
         let mask = x86_64::_mm256_and_ps(mask, one);
         let fx = x86_64::_mm256_sub_ps(tmp, mask);
 
-        let c1 = x86_64::_mm256_set1_ps(CEPHES_EXP_C1);
+        let c1 = x86_64::_mm256_set1_ps(CEPHES_EXP_C1 as f32);
         let tmp = x86_64::_mm256_mul_ps(fx, c1);
-        let c2 = x86_64::_mm256_set1_ps(CEPHES_EXP_C2);
+        let c2 = x86_64::_mm256_set1_ps(CEPHES_EXP_C2 as f32);
         let z4 = x86_64::_mm256_mul_ps(fx, c2);
         let x4 = x86_64::_mm256_sub_ps(x4, tmp);
         let x4 = x86_64::_mm256_sub_ps(x4, z4);
 
         let z4 = x86_64::_mm256_mul_ps(x4, x4);
 
-        let y4 = x86_64::_mm256_set1_ps(CEPHES_EXP_P0);
+        let y4 = x86_64::_mm256_set1_ps(CEPHES_EXP_P0 as f32);
         let y4 = x86_64::_mm256_mul_ps(y4, x4);
-        let exp_p1 = x86_64::_mm256_set1_ps(CEPHES_EXP_P1);
+        let exp_p1 = x86_64::_mm256_set1_ps(CEPHES_EXP_P1 as f32);
         let y4 = x86_64::_mm256_add_ps(y4, exp_p1);
         let y4 = x86_64::_mm256_mul_ps(y4, x4);
-        let exp_p2 = x86_64::_mm256_set1_ps(CEPHES_EXP_P2);
+        let exp_p2 = x86_64::_mm256_set1_ps(CEPHES_EXP_P2 as f32);
         let y4 = x86_64::_mm256_add_ps(y4, exp_p2);
         let y4 = x86_64::_mm256_mul_ps(y4, x4);
-        let exp_p3 = x86_64::_mm256_set1_ps(CEPHES_EXP_P3);
+        let exp_p3 = x86_64::_mm256_set1_ps(CEPHES_EXP_P3 as f32);
         let y4 = x86_64::_mm256_add_ps(y4, exp_p3);
         let y4 = x86_64::_mm256_mul_ps(y4, x4);
-        let exp_p4 = x86_64::_mm256_set1_ps(CEPHES_EXP_P4);
+        let exp_p4 = x86_64::_mm256_set1_ps(CEPHES_EXP_P4 as f32);
         let y4 = x86_64::_mm256_add_ps(y4, exp_p4);
         let y4 = x86_64::_mm256_mul_ps(y4, x4);
-        let exp_p5 = x86_64::_mm256_set1_ps(CEPHES_EXP_P5);
+        let exp_p5 = x86_64::_mm256_set1_ps(CEPHES_EXP_P5 as f32);
         let y4 = x86_64::_mm256_add_ps(y4, exp_p5);
         let y4 = x86_64::_mm256_mul_ps(y4, z4);
         let y4 = x86_64::_mm256_add_ps(y4, x4);
